chore:renamed demo file
